e2e_tests: Add ECDSA key load test in new module
This patch moves the tests related to keys to a separate module.
Also adds a test to load an ECDSA key from the parsec provider.
