+++ parsec-openssl-provider-shared/e2e_tests/tests/mod.rs
@@ -1,5 +1,6 @@
 // Copyright 2024 Contributors to the Parsec project.
 // SPDX-License-Identifier: Apache-2.0
 
+mod keys;
 mod provider;
 mod sign;
